+++ CHANGELOG.md
@@ -8,7 +8,11 @@ All notable changes to the "VSCode SystemVerilog - Language Support" extension w
 - Back-end Language server for Systemverilog
 - Complete syntax highlighting
 
-## [0.0.3] - 2018-02-03
+## [0.2.4] - 2018-02-06
+- DocumentSymbolProvider now provides instatiated modules
+- Adjusted TM of module instantiation 
+
+## [0.2.3] - 2018-02-03
 - Added extension icon
 - Updated description